requirements.txt
flask
torch
torchvision
xgboost
pillow
numpy
joblib
gunicorn